requirements.txt
numpy
packaging
pandas
python-dotenv
python-engineio
python-socketio
websocket-client
uvicorn
fastapi
pandas-ta
numba
httpx
urllib3
requests
six >= 1.10
python_dateutil >= 2.5.3
setuptools >= 21.0.0
firebase_admin